requirements.txt
torch==2.2.1
torchvision==0.17.1
tensorflow==2.15.0
tensorflow-datasets==4.9.3
transformers==4.38.2
diffusers==0.27.0
imageio==2.9.0
imageio-ffmpeg==0.4.4
pillow==10.0.0
numpy==1.24.4
transformers==4.38.2
accelerate==0.29.3
safetensors==0.4.2
tensorboard==2.14.0
imageio==2.34.0
imageio-ffmpeg==0.4.9
pillow==10.0.0
numpy==1.24.4
diffusers